pyproject.toml
[project]
name = "logai"
version = "0.1.0"
description = ""
authors = [
    {name = "Arthur Gonzaga",email = "arthursgonzaga@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=1.76.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "logai", from = "src"}]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
